requirements.txt
six
PyYaml

# for loading yaml into ordered dicts
yamlloader

# TODO: we can likely eliminate this
jinja2
